+++ tests/cypress/tsconfig.json
@@ -2,7 +2,10 @@
   "compilerOptions": {
     "target": "es5",
     "lib": ["es5", "dom"],
-    "types": ["cypress", "node"]
+    "types": ["cypress", "node"],
+    "paths": {
+      "~/*": ["./*"]
+    }
   },
   "include": ["**/*.ts"]
 }
